Make benchmarks more uniform. Fix issues following pooling pattern.
